requirements.txt
# AI/ML Dependencies
openai==1.12.0
pinecone-client==3.0.3
langchain==0.1.16
langchain-openai==0.1.1
langchain-community==0.0.32
langchain-text-splitters==0.0.1
tiktoken==0.6.0

# Web Framework
flask==3.0.3
flask-cors==4.0.0
flask-limiter==3.5.0
gunicorn==21.2.0

# Data Processing
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
typing-extensions==4.12.2

# HTTP & API
requests==2.31.0
httpx==0.27.0

# Monitoring & Logging
prometheus-client==0.20.0
structlog==24.1.0

# Development & Testing
pytest==8.1.1
pytest-cov==5.0.0
pytest-asyncio==0.23.5
black==24.3.0
flake8==7.0.0
isort==5.13.2

# Utilities
python-multipart==0.0.9
Jinja2==3.1.3
markupsafe==2.1.5